+++ ZAPD/ZScalar.cpp
@@ -19,7 +19,7 @@ ZScalar::ZScalar(const ZScalarType scalarType, ZFile* nParent) : ZScalar(nParent
 }
 
 void ZScalar::ExtractFromXML(tinyxml2::XMLElement* reader, const std::vector<uint8_t>& nRawData,
-                             const int32_t nRawDataIndex, const std::string& nRelPath)
+                             const uint32_t nRawDataIndex, const std::string& nRelPath)
 {
 	ZResource::ExtractFromXML(reader, nRawData, nRawDataIndex, nRelPath);
 }
@@ -107,7 +107,7 @@ std::string ZScalar::MapScalarTypeToOutputType(const ZScalarType scalarType)
 	}
 }
 
-int32_t ZScalar::MapTypeToSize(const ZScalarType scalarType)
+size_t ZScalar::MapTypeToSize(const ZScalarType scalarType)
 {
 	switch (scalarType)
 	{
@@ -146,7 +146,7 @@ void ZScalar::ParseRawData()
 	ZScalar::ParseRawData(rawData, rawDataIndex);
 }
 
-void ZScalar::ParseRawData(const std::vector<uint8_t>& data, const int32_t offset)
+void ZScalar::ParseRawData(const std::vector<uint8_t>& data, const uint32_t offset)
 {
 	switch (scalarType)
 	{
