+++ ZAPD/ZScalar.h
@@ -56,11 +56,11 @@ public:
 	void ParseRawData() override;
 
 	void ExtractFromXML(tinyxml2::XMLElement* reader, const std::vector<uint8_t>& nRawData,
-	                    const int32_t nRawDataIndex, const std::string& nRelPath) override;
-	static int32_t MapTypeToSize(const ZScalarType scalarType);
+	                    const uint32_t nRawDataIndex, const std::string& nRelPath) override;
+	static size_t MapTypeToSize(const ZScalarType scalarType);
 	static ZScalarType MapOutputTypeToScalarType(const std::string& type);
 	static std::string MapScalarTypeToOutputType(const ZScalarType scalarType);
 
 protected:
-	void ParseRawData(const std::vector<uint8_t>& data, const int32_t offset);
+	void ParseRawData(const std::vector<uint8_t>& data, const uint32_t offset);
 };
